+++ python/pathway/stdlib/stateful/__init__.py
@@ -1,0 +1,9 @@
+# Copyright © 2023 Pathway
+
+from __future__ import annotations
+
+from .deduplicate import deduplicate
+
+__all__ = [
+    "deduplicate",
+]