+++ tests/test_jsonlines.rs
@@ -8,6 +8,7 @@ use std::sync::Arc;
 
 use pathway_engine::connectors::data_format::{JsonLinesParser, ParsedEvent};
 use pathway_engine::connectors::data_storage::{ConnectorMode, FilesystemReader, ReadMethod};
+use pathway_engine::connectors::SessionType;
 use pathway_engine::engine::Value;
 
 #[test]
@@ -25,6 +26,7 @@ fn test_jsonlines_ok() -> eyre::Result<()> {
         HashMap::new(),
         true,
         HashMap::new(),
+        SessionType::Native,
     );
 
     let entries = read_data_from_reader(Box::new(reader), Box::new(parser))?;
@@ -64,6 +66,7 @@ fn test_jsonlines_incorrect_key() -> eyre::Result<()> {
         HashMap::new(),
         true,
         HashMap::new(),
+        SessionType::Native,
     );
 
     assert_error_shown(
@@ -90,6 +93,7 @@ fn test_jsonlines_incomplete_key_to_null() -> eyre::Result<()> {
         HashMap::new(),
         false,
         HashMap::new(),
+        SessionType::Native,
     );
 
     let entries = read_data_from_reader(Box::new(reader), Box::new(parser))?;
@@ -113,6 +117,7 @@ fn test_jsonlines_incorrect_values() -> eyre::Result<()> {
         HashMap::new(),
         true,
         HashMap::new(),
+        SessionType::Native,
     );
 
     assert_error_shown(
@@ -147,6 +152,7 @@ fn test_jsonlines_types_parsing() -> eyre::Result<()> {
         HashMap::new(),
         true,
         HashMap::new(),
+        SessionType::Native,
     );
 
     let entries = read_data_from_reader(Box::new(reader), Box::new(parser))?;
@@ -201,6 +207,7 @@ fn test_jsonlines_complex_paths() -> eyre::Result<()> {
         routes,
         true,
         HashMap::new(),
+        SessionType::Native,
     );
 
     let entries = read_data_from_reader(Box::new(reader), Box::new(parser))?;
@@ -260,6 +267,7 @@ fn test_jsonlines_complex_paths_error() -> eyre::Result<()> {
         routes,
         true,
         HashMap::new(),
+        SessionType::Native,
     );
 
     assert_error_shown(
@@ -301,6 +309,7 @@ fn test_jsonlines_complex_path_ignore_errors() -> eyre::Result<()> {
         routes,
         false,
         HashMap::new(),
+        SessionType::Native,
     );
 
     let entries = read_data_from_reader(Box::new(reader), Box::new(parser))?;
@@ -324,6 +333,7 @@ fn test_jsonlines_incorrect_key_verbose_error() -> eyre::Result<()> {
         HashMap::new(),
         true,
         HashMap::new(),
+        SessionType::Native,
     );
 
     assert_error_shown(
@@ -353,6 +363,7 @@ fn test_jsonlines_incorrect_jsonpointer_verbose_error() -> eyre::Result<()> {
         routes,
         true,
         HashMap::new(),
+        SessionType::Native,
     );
 
     assert_error_shown(
@@ -379,6 +390,7 @@ fn test_jsonlines_failed_to_parse_field() -> eyre::Result<()> {
         HashMap::new(),
         true,
         HashMap::new(),
+        SessionType::Native,
     );
 
     assert_error_shown(